Added Seeking
Tear's now follow you around. Still version 1 of the movement. Added
wrapping, and used the pythagorean theorem to get the distance, then
angle. I used that angle to calculate where the tear should be headed,
and then multiplied it by it's speed.

Also added the top_speed variable for setting the current top_speed of
the collectible.

Also added the randomize function, to not collect, just reassign
location to a collectible.