Change Ship Acceleration and Add Elapsed Time
The acceleration method is now slightly different. Instead of adding a
single flat rate, the new way it works is by checking to see if the
current speed is less or more than half of MAX_SPEED. If it's over half
and you are trying to slow down it does so in greater increment,
likewise if you are trying to speed up from less than half speed it
speeds more quickly.

Also added an elapsed timer, always counting up by (1/60) of a second
per animation frame. This allows for me to do things like, instead of
just having score thresholds, also add timer thresholds. This way if the
player sits and scores nothing but time for too long the game continues.