Gem Spawning and Minor Changes
Change how gem spawning works. It was a big overhaul in code, so now
it's more robust, but it works practically the same way. The gems are a
spawned as a function of the score.
Blue gems are now here.
Pink gems, in their most basic form, are now present and flying.
Move the title around a bit.
Add in the images for some new stuff.