Fix a bug where images wouldn't display in Safari
There was previously a problem where images wouldn't display in safari.
This was caused by a bug where I was trying to load all of the images
once every 1/60th of a second. In most browsers it cached the image, but
in Safari it didn't.
Now I preload all of the images and sounds and pass them through as
necessary.